refactor(routes): refactor router and rename routes folder to pages
