+++ src/pages/battle/index.js
