+++ src/pages/battles/Battles.js
