+++ src/pages/battles/index.js
