+++ src/pages/editor/index.js
