+++ src/pages/kuski/PlayedBattles.css
