+++ tools/render.js
@@ -23,11 +23,6 @@ import runServer from './runServer';
 // ];
 const routes = [
   '/',
-  '/contact',
-  '/login',
-  '/register',
-  '/about',
-  '/privacy',
   '/404', // https://help.github.com/articles/creating-a-custom-404-page-for-your-github-pages-site/
 ];
